+++ CHANGELOG.md
@@ -16,6 +16,8 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 - The local authority contact in exports is now the "main" contact as selected
   in the UI. If none is selected, then the exported contact is the next contact
   with the category of "local authority"
+- Users are now asked which type of project to add with an explanation of each
+  type.
 
 ### Added